fixes based on review
