+++ pkg/cache/kubeconfig.go
@@ -8,8 +8,6 @@ import (
 	"time"
 
 	"github.com/jellydator/ttlcache/v3"
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promauto"
 	"go.uber.org/zap"
 	corev1 "k8s.io/api/core/v1"
 	k8serr "k8s.io/apimachinery/pkg/api/errors"
@@ -27,14 +25,6 @@ var (
 		ttlcache.WithTTL[string, string](ttl),
 		ttlcache.WithDisableTouchOnHit[string, string](),
 	)
-
-	cacheSizeMetric = promauto.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Namespace: "kmc",
-			Subsystem: "kubeconfig_cache",
-			Name:      "size",
-			Help:      "Number of items in the kubeconfig cache.",
-		}, nil)
 )
 
 // GetKubeConfigFromCache returns the kubeconfig from the cache if it is not expired.
@@ -70,10 +60,6 @@ func GetKubeConfigFromCache(logger *zap.SugaredLogger, coreV1 v1.CoreV1Interface
 	}
 
 	return kubeConfig, err
-}
-
-func recordMetrics() {
-	cacheSizeMetric.With(prometheus.Labels{}).Set(float64(kubeConfigCache.Len()))
 }
 
 // getkubeConfigFromSecret gets the kubeconfig from the secret.
